refactor: move logic to another file
